test_app/requirements.txt
docxcompose==1.3.5
docxtpl==0.16.0
editpyxl==0.1.10
et-xmlfile==1.1.0
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
python-docx==0.8.11
six==1.16.0
xlrd==2.0.1
python-dateutil~=2.8.2